requirements.txt
altgraph==0.17.4
cx_Freeze==7.2.7
cx_Logging==3.2.1
defusedxml==0.7.1
et_xmlfile==2.0.0
lief==0.15.1
lml==0.1.0
lxml==5.3.0
Nuitka==2.5.4
numpy==2.1.3
odfpy==1.4.1
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pyexcel-io==0.6.7
pyexcel-ods==0.6.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyoxidizer==0.24.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
pywin32-ctypes==0.2.3
setuptools==75.6.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
zstandard==0.23.0
